So here's a list of new changes (The last being the most important I think): Added new icons to give application a bit more visual appeal,also changed the executable file icon to a key image
Now KeyVal checks for out-of-range input Keys in the Database and outputs error message without crashing the program

The KeyVal Editing window is now ordered by the Keys column number where it used to be ordered by an unseen id

When the program starts up the keyboard cursor will start in the first key input box.

Logging is added to a text file: KeyValLog.txt

A new window to change the directory path has been added under the Edit Menu
I did this as a work around for the Parallels drive mapping issue. I discovered that when Parallels is running it uses a "psf" (parallels shared folder) to map to the mac desktop and directories into windows. Because of how this is being done, windows interprets it as a network path and for some reason won't connect/access with write privileges and therefore crashes the program. By simply removing "psf/home/" to a drive letter i.e. "Z:/" the program will work. You can find the parallels Drive Letter under 'My Computer'. Please see the attached image.

You could also change this to point to another KeyVal.s3db database located in another folder or directory (say a dropbox shared drive or another existing database etc.) if you’d like as long as long as you provide the full path i.e "C:/Users/Mike/My Documents/KeVal.s3db".

If the program gives an error saying it can’t locate the database open the "Change Directory Directory Path" window and it should default to the KeyVal.s3db database within the parent folder when you close the window.